+++ packages/preact-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/preact-query-persist-client
 
+## 5.101.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.101.2
+  - @tanstack/query-persist-client-core@5.101.2
+
 ## 5.101.1
 
 ### Patch Changes