+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.1.36
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.101.2
+  - @tanstack/svelte-query@6.1.36
+
 ## 6.1.35
 
 ### Patch Changes